+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/properties/events/MPROnHit.java
@@ -52,11 +52,6 @@ public class MPROnHit extends MPREvent {
 		if (this.target == null)
 			throw new JsonValidationException("Missing \"target\" for OnHit object: %s".formatted(this));
 
-		if (this.potionEffects != null) {
-			for (MPRPotionEffect potionEffect : this.potionEffects)
-				potionEffect.validate();
-		}
-
 		if (this.damageType != null) {
 			if (this.damageType.startsWith("#")) {
 				ResourceLocation rl = ResourceLocation.tryParse(this.damageType.substring(1));